+++ lib/src/platform/image_loader_mobile.dart
@@ -13,6 +13,7 @@ class PlatformImageLoader {
     bool enableCache = true,
     List<ImageLoadStrategy>? strategies,
     ImageLoadCallback? onStrategyResolved,
+    bool preventNativeInteraction = true,
   }) async {
     return Image.network(
       url,